+++ start/src/App.vue
@@ -1,41 +1,23 @@
-<script setup lang="ts">
+<script lang="ts" setup>
   import { ref } from 'vue'
 
-  const visible1 = ref(false)
+  const url = ref(
+    'https://img1.baidu.com/it/u=1016138010,1907110459&fm=253&fmt=auto&app=138&f=JPEG?w=500&h=500'
+  )
+
+  const change = (): void => {
+    url.value =
+      'https://img2.baidu.com/it/u=3618236253,1028428296&fm=253&app=138&size=w931&n=0&f=JPEG&fmt=auto?sec=1683392400&t=f243a7fe228993127dbd7ac56bfbef3f'
+  }
 </script>
 
 <template>
-  <f-button type="primary" @click="visible1 = true">打开</f-button>
+  <f-image width="200px" :src="url" />
+  <f-avatar width="200px" :src="url" />
+  <img width="20" :src="url" />
+  <img width="200" :src="url" />
 
-  <f-dialog v-model:visible="visible1" title="标题文字">
-    这是一个 Dialog
-    <h1>1</h1>
-    <h1>1</h1>
-    <h1>1</h1>
-    <h1>1</h1>
-    <h1>1</h1>
-    <h1>1</h1>
-    <h1>1</h1>
-    <h1>1</h1>
-    <h1>1</h1>
-    <h1>1</h1>
-    <h1>1</h1>
-    <h1>1</h1>
-    <h1>1</h1>
-    <h1>1</h1>
-    <h1>1</h1>
-    <h1>1</h1>
-    <h1>1</h1>
-    <h1>1</h1>
-    <h1>1</h1>
-    <h1>1</h1>
-    <h1>1</h1>
-    <h1>1</h1>
-    <h1>1</h1>
+  <button>123</button>
 
-    <template #footer>
-      <f-button type="default">默认按钮</f-button>
-      <f-button type="primary">主要按钮</f-button>
-    </template>
-  </f-dialog>
+  <f-button :on-click="change">切换</f-button>大大的
 </template>